move send batch txs logic to new file
