+++ bench-tps/src/lib.rs
@@ -4,3 +4,4 @@ pub mod bench_tps_client;
 pub mod cli;
 pub mod keypairs;
 mod perf_utils;
+pub mod send_batch;
